requirements.txt
django==3.1
psycopg2-binary
Pillow
djangorestframework
python-decouple
django-phonenumber-field
phonenumbers